Fix error: Caused by: java.lang.ClassCastException: max.hubbard.bettershops.Shops.Items.FileShopItem cannot be cast to java.lang.Integer

+++ src/main/java/max/hubbard/bettershops/Menus/ShopMenus/MainBuying.java
@@ -300,7 +300,7 @@ public class MainBuying implements ShopMenu {
                     if (!it.isSelling()) {
 
                         if (obj[0] != null) {
-                            if (it.getPrice() != (int) obj[0]) {
+                            if (it.getPrice() != (double) obj[0]) {
                                 continue;
                             }
                         }
